requirements.txt
autopep8==1.4.4
Django==2.2.7
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
Pillow==6.2.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
